[#378]; refactor: masterdata/support의 initialization 폴더를 initializer로 이름 변경
- 클래스 접미사(*Initializer)와 폴더명을 맞춤 (exception, configuration 폴더 컨벤션과 통일)
